# [PERF] core: speed up file_open