Description of the issue/feature this PR addresses:
Checking that a path exists right before opening it is useless because `open` will already check if the path is a readable file.
We can already reduce a few syscalls during startup in `odoo.tools.config` and when reading views from the disk.

Current behavior before PR:
- file_open call
  - file_path
    - exists
  - is_file
  - open

Desired behavior after PR is merged:
- file_open call
  - file_path
    - exists (only for relative paths where module is not loaded)
  - open

See details in commit messages.

## benchmark

```
# before
In [4]: %timeit file_open('base/models/res_partner.py')
27.1 μs ± 246 ns per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
In [5]: %timeit file_open('/opt/odoo/odoo/addons/base/models/res_partner.py')
26.2 μs ± 316 ns per loop (mean ± std. dev. of 7 runs, 10,000 loops each)

# after
In [3]: %timeit file_open('base/models/res_partner.py')
11.4 μs ± 48.5 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)
In [4]: %timeit file_open('/opt/odoo/odoo/addons/base/models/res_partner.py')
12.6 μs ± 119 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)
```

---
I confirm I have signed the CLA and read the PR guidelines at www.odoo.com/submit-pr