@Julien00859 2025-05-27
@kmagusiak benchmark and tag should be `[PERF]`