@kmagusiak 2025-05-27
> @kmagusiak benchmark and tag should be `[PERF]`

Updated description and commit messages.