@rco-odoo 2025-06-10
@robodoo override=ci/security

The changed code is the implementation of the "safe" API of Odoo to access the filesystem.
* `os.path`, `os.sep` and `open` are used as before, and always prevent forbidden accesses;
* `sys.modules` is used to check that you don't access some file outside of modules.